requirements.txt
streamlit>=1.32,<2.0
pandas>=2.1,<3.0
numpy>=1.26,<3.0
requests>=2.31,<3.0
openai>=1.30,<2.0
python-dotenv>=1.0,<2.0
beautifulsoup4>=4.12,<5.0
lxml>=4.9,<6.0
openpyxl>=3.1,<4.0
python-docx>=1.1,<1.2
altair>=5.0,<6.0
sqlalchemy==2.0.32
serpapi==0.1.5
python-dotenv==1.0.1

